+++ app/src/main/java/com/example/weatherapp/di/ApplicationModule.kt
@@ -17,12 +17,12 @@ import com.example.weatherapp.data.userpreference.UserPreferenceSerializer
 import com.example.weatherapp.data.weatherforecast.repository.WeatherForecastAPIRepositoryImpl
 import com.example.weatherapp.data.weatherforecast.repository.base.WeatherForecastRepository
 import com.example.weatherapp.data.weatherforecast.service.WeatherForecastService
-import com.example.weatherapp.usecase.currentweather.GetCurrentWeatherUseCaseImpl
-import com.example.weatherapp.usecase.currentweather.base.GetCurrentWeatherUseCase
-import com.example.weatherapp.usecase.geocoding.GetPlacesForSearchQueryUseCaseImpl
-import com.example.weatherapp.usecase.geocoding.base.GetPlacesForSearchQueryUseCase
-import com.example.weatherapp.usecase.weatherforecast.GetWeatherForecastForLocationUseCaseImpl
-import com.example.weatherapp.usecase.weatherforecast.base.GetWeatherForecastForLocationUseCase
+import com.example.weatherapp.domain.currentweather.GetCurrentWeatherUseCaseImpl
+import com.example.weatherapp.domain.currentweather.base.GetCurrentWeatherUseCase
+import com.example.weatherapp.domain.geocoding.GetPlacesForSearchQueryUseCaseImpl
+import com.example.weatherapp.domain.geocoding.base.GetPlacesForSearchQueryUseCase
+import com.example.weatherapp.domain.weatherforecast.GetWeatherForecastForLocationUseCaseImpl
+import com.example.weatherapp.domain.weatherforecast.base.GetWeatherForecastForLocationUseCase
 import com.squareup.moshi.Moshi
 import dagger.Binds
 import dagger.Module
