+++ app/src/main/java/com/example/weatherapp/domain/weatherforecast/GetWeatherForecastForLocationUseCaseImpl.kt
@@ -1,9 +1,9 @@
-package com.example.weatherapp.usecase.weatherforecast
+package com.example.weatherapp.domain.weatherforecast
 
 import com.example.weatherapp.data.common.Result
 import com.example.weatherapp.data.weatherforecast.model.ForecastAPIResponse
 import com.example.weatherapp.data.weatherforecast.repository.base.WeatherForecastRepository
-import com.example.weatherapp.usecase.weatherforecast.base.GetWeatherForecastForLocationUseCase
+import com.example.weatherapp.domain.weatherforecast.base.GetWeatherForecastForLocationUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
